feat: add sidebar component and integrate
