+++ components/results-display.tsx
@@ -16,13 +16,15 @@ import {
   AlertTriangle,
   HelpCircle,
   Package,
-  Target,
   Save,
 } from "lucide-react"
 import type { GWPCalculation } from "@/lib/gwp-calculator"
 import { projectsService } from "@/lib/services/projects-service"
 import { useAppState } from "@/lib/services/app-state"
 import { notificationService } from "@/lib/services/notification-service"
+
+import { MacroGroupChart } from "./results-display/macro-group-chart"
+import { BreakdownChart } from "./results-display/breakdown-chart"
 
 interface ResultsDisplayProps {
   gwpResults: GWPCalculation
@@ -353,132 +355,12 @@ export default function ResultsDisplay({ gwpResults, onReset }: ResultsDisplayPr
         </TabsContent>
 
         <TabsContent value="macrogroups" className="space-y-6">
-          {/* Macro-group impact visualization */}
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Package className="h-5 w-5" />
-                PCR Macro-Group Impact Analysis
-              </CardTitle>
-              <CardDescription>
-                Visual representation of how each PCR macro-group contributes to the total GWP
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-6">
-                {/* Visual pie chart representation */}
-                <div className="flex justify-center">
-                  <div className="relative w-64 h-64">
-                    <svg viewBox="0 0 200 200" className="w-full h-full">
-                      {macroGroupAnalysis.map((group, index) => {
-                        const startAngle = macroGroupAnalysis
-                          .slice(0, index)
-                          .reduce((sum, g) => sum + (g.percentage / 100) * 360, 0)
-                        const endAngle = startAngle + (group.percentage / 100) * 360
-                        const largeArcFlag = group.percentage > 50 ? 1 : 0
-
-                        const x1 = 100 + 80 * Math.cos((startAngle * Math.PI) / 180)
-                        const y1 = 100 + 80 * Math.sin((startAngle * Math.PI) / 180)
-                        const x2 = 100 + 80 * Math.cos((endAngle * Math.PI) / 180)
-                        const y2 = 100 + 80 * Math.sin((endAngle * Math.PI) / 180)
-
-                        const pathData = [
-                          `M 100 100`,
-                          `L ${x1} ${y1}`,
-                          `A 80 80 0 ${largeArcFlag} 1 ${x2} ${y2}`,
-                          `Z`,
-                        ].join(" ")
-
-                        const colors = ["#3B82F6", "#10B981", "#F59E0B", "#EF4444", "#8B5CF6", "#6366F1", "#EC4899"]
-
-                        return (
-                          <path
-                            key={group.id}
-                            d={pathData}
-                            fill={colors[index % colors.length]}
-                            stroke="white"
-                            strokeWidth="2"
-                            className="hover:opacity-80 transition-opacity"
-                          />
-                        )
-                      })}
-                    </svg>
-                  </div>
-                </div>
-
-                {/* Legend and details */}
-                <div className="grid md:grid-cols-2 gap-4">
-                  {macroGroupAnalysis.map((group, index) => (
-                    <Card key={group.id} className="relative overflow-hidden">
-                      <div className={`absolute left-0 top-0 bottom-0 w-1 ${getColorForIndex(index)}`} />
-                      <CardContent className="p-4 pl-6">
-                        <div className="flex items-center justify-between mb-2">
-                          <div>
-                            <div className="flex items-center gap-2">
-                              <span className="font-mono text-sm bg-gray-100 px-2 py-1 rounded">{group.code}</span>
-                              <span className="font-medium text-sm">{group.name}</span>
-                            </div>
-                          </div>
-                          <div className="text-right">
-                            <div className={`text-lg font-bold ${getTextColorForIndex(index)}`}>
-                              {formatNumber(group.totalGWP / 1000)}t
-                            </div>
-                            <div className="text-xs text-gray-500">{group.percentage.toFixed(1)}% of total</div>
-                          </div>
-                        </div>
-
-                        {/* Progress bar */}
-                        <div className="w-full bg-gray-200 rounded-full h-2 mb-2">
-                          <div
-                            className={`h-2 rounded-full ${getColorForIndex(index)}`}
-                            style={{ width: `${Math.min(100, group.percentage)}%` }}
-                          />
-                        </div>
-
-                        <div className="text-xs text-gray-600">
-                          {group.materials.length} materials • Avg:{" "}
-                          {formatNumber(group.totalGWP / group.materials.length / 1000)}t per material
-                        </div>
-                      </CardContent>
-                    </Card>
-                  ))}
-                </div>
-
-                {/* Impact ranking */}
-                <Card>
-                  <CardHeader>
-                    <CardTitle className="flex items-center gap-2">
-                      <Target className="h-5 w-5" />
-                      Impact Ranking
-                    </CardTitle>
-                  </CardHeader>
-                  <CardContent>
-                    <div className="space-y-3">
-                      {macroGroupAnalysis.map((group, index) => (
-                        <div key={group.id} className="flex items-center gap-4">
-                          <div className="flex items-center justify-center w-8 h-8 rounded-full bg-gray-100 text-sm font-bold">
-                            {index + 1}
-                          </div>
-                          <div className="flex-1">
-                            <div className="flex items-center gap-2">
-                              <span className="font-mono text-xs bg-gray-100 px-2 py-1 rounded">{group.code}</span>
-                              <span className="font-medium">{group.name}</span>
-                            </div>
-                          </div>
-                          <div className="text-right">
-                            <div className={`font-bold ${getTextColorForIndex(index)}`}>
-                              {formatNumber(group.totalGWP / 1000)}t CO₂eq
-                            </div>
-                            <div className="text-xs text-gray-500">{group.percentage.toFixed(1)}%</div>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  </CardContent>
-                </Card>
-              </div>
-            </CardContent>
-          </Card>
+          <MacroGroupChart
+            data={macroGroupAnalysis}
+            formatNumber={formatNumber}
+            getColorForIndex={getColorForIndex}
+            getTextColorForIndex={getTextColorForIndex}
+          />
         </TabsContent>
 
         <TabsContent value="materials" className="space-y-6">
@@ -530,69 +412,7 @@ export default function ResultsDisplay({ gwpResults, onReset }: ResultsDisplayPr
         </TabsContent>
 
         <TabsContent value="breakdown" className="space-y-6">
-          <Card>
-            <CardHeader>
-              <CardTitle>Emissions Breakdown by Phase</CardTitle>
-              <CardDescription>Distribution of emissions across different lifecycle phases</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-6">
-                <div className="grid md:grid-cols-3 gap-4">
-                  <div className="text-center p-4 border rounded-lg">
-                    <p className="text-2xl font-bold text-blue-600">
-                      {formatNumber(gwpResults.breakdown.production / 1000)}
-                    </p>
-                    <p className="text-sm text-gray-600">Material Production</p>
-                    <p className="text-xs text-gray-500">75% of total</p>
-                  </div>
-                  <div className="text-center p-4 border rounded-lg">
-                    <p className="text-2xl font-bold text-orange-600">
-                      {formatNumber(gwpResults.breakdown.transport / 1000)}
-                    </p>
-                    <p className="text-sm text-gray-600">Transportation</p>
-                    <p className="text-xs text-gray-500">15% of total</p>
-                  </div>
-                  <div className="text-center p-4 border rounded-lg">
-                    <p className="text-2xl font-bold text-purple-600">
-                      {formatNumber(gwpResults.breakdown.processing / 1000)}
-                    </p>
-                    <p className="text-sm text-gray-600">Processing</p>
-                    <p className="text-xs text-gray-500">10% of total</p>
-                  </div>
-                </div>
-
-                <div className="space-y-3">
-                  <div>
-                    <div className="flex justify-between mb-1">
-                      <span className="text-sm font-medium">Material Production</span>
-                      <span className="text-sm text-gray-600">75%</span>
-                    </div>
-                    <div className="w-full bg-gray-200 rounded-full h-3">
-                      <div className="bg-blue-600 h-3 rounded-full" style={{ width: "75%" }} />
-                    </div>
-                  </div>
-                  <div>
-                    <div className="flex justify-between mb-1">
-                      <span className="text-sm font-medium">Transportation</span>
-                      <span className="text-sm text-gray-600">15%</span>
-                    </div>
-                    <div className="w-full bg-gray-200 rounded-full h-3">
-                      <div className="bg-orange-600 h-3 rounded-full" style={{ width: "15%" }} />
-                    </div>
-                  </div>
-                  <div>
-                    <div className="flex justify-between mb-1">
-                      <span className="text-sm font-medium">Processing</span>
-                      <span className="text-sm text-gray-600">10%</span>
-                    </div>
-                    <div className="w-full bg-gray-200 rounded-full h-3">
-                      <div className="bg-purple-600 h-3 rounded-full" style={{ width: "10%" }} />
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
+          <BreakdownChart data={gwpResults.breakdown} formatNumber={formatNumber} />
         </TabsContent>
 
         <TabsContent value="recommendations" className="space-y-6">
